Split up dbt.contracts.results into various dbt.artifacts files
