+++ core/dbt/cli/main.py
@@ -17,10 +17,8 @@ from dbt.cli.exceptions import (
     DbtUsageException,
 )
 from dbt.contracts.graph.manifest import Manifest
-from dbt.contracts.results import (
-    CatalogArtifact,
-    RunExecutionResult,
-)
+from dbt.artifacts.catalog import CatalogArtifact
+from dbt.artifacts.run import RunExecutionResult
 from dbt.common.events.base_types import EventMsg
 from dbt.task.build import BuildTask
 from dbt.task.clean import CleanTask
