+++ core/dbt/contracts/sql.py
@@ -6,13 +6,8 @@ from typing import Optional, List, Any, Dict, Sequence
 from dbt.common.dataclass_schema import dbtClassMixin
 
 from dbt.contracts.graph.nodes import ResultNode
-from dbt.contracts.results import (
-    RunResult,
-    RunResultsArtifact,
-    TimingInfo,
-    ExecutionResult,
-    RunExecutionResult,
-)
+from dbt.artifacts.results import TimingInfo, ExecutionResult
+from dbt.artifacts.run import RunResult, RunResultsArtifact, RunExecutionResult
 from dbt.artifacts.base import VersionedSchema, schema_version
 from dbt.logger import LogMessage
