+++ core/dbt/contracts/state.py
@@ -2,8 +2,8 @@ from pathlib import Path
 from typing import Optional
 
 from dbt.contracts.graph.manifest import WritableManifest
-from dbt.contracts.results import FreshnessExecutionResultArtifact
-from dbt.contracts.results import RunResultsArtifact
+from dbt.artifacts.freshness import FreshnessExecutionResultArtifact
+from dbt.artifacts.run import RunResultsArtifact
 from dbt.common.events.functions import fire_event
 from dbt.events.types import WarnStateTargetEqual
 from dbt.exceptions import IncompatibleSchemaError
