+++ core/dbt/task/base.py
@@ -16,14 +16,9 @@ from dbt.config import RuntimeConfig, Project
 from dbt.config.profile import read_profile
 from dbt.constants import DBT_PROJECT_FILE_NAME
 from dbt.contracts.graph.manifest import Manifest
-from dbt.contracts.results import (
-    NodeStatus,
-    RunResult,
-    collect_timing_info,
-    RunStatus,
-    RunningStatus,
-    TimingInfo,
-)
+from dbt.artifacts.results import TimingInfo, collect_timing_info
+from dbt.artifacts.results import NodeStatus, RunningStatus, RunStatus
+from dbt.artifacts.run import RunResult
 from dbt.common.events.contextvars import get_node_info
 from dbt.common.events.functions import fire_event
 from dbt.events.types import (
