+++ core/dbt/task/build.py
@@ -5,13 +5,13 @@ from .snapshot import SnapshotRunner as snapshot_model_runner
 from .seed import SeedRunner as seed_runner
 from .test import TestRunner as test_runner
 
-from dbt.contracts.results import NodeStatus
+from dbt.artifacts.results import NodeStatus, RunStatus
+from dbt.artifacts.run import RunResult
 from dbt.common.exceptions import DbtInternalError
 from dbt.graph import ResourceTypeSelector
 from dbt.node_types import NodeType
 from dbt.task.test import TestSelector
 from dbt.task.base import BaseRunner
-from dbt.contracts.results import RunResult, RunStatus
 from dbt.common.events.functions import fire_event
 from dbt.events.types import LogStartLine, LogModelResult
 from dbt.common.events.base_types import EventLevel
