+++ core/dbt/task/clone.py
@@ -4,7 +4,7 @@ from typing import AbstractSet, Any, List, Iterable, Set
 from dbt.adapters.base import BaseRelation
 from dbt.clients.jinja import MacroGenerator
 from dbt.context.providers import generate_runtime_model_context
-from dbt.contracts.results import RunStatus, RunResult
+from dbt.artifacts.run import RunStatus, RunResult
 from dbt.common.dataclass_schema import dbtClassMixin
 from dbt.common.exceptions import DbtInternalError, CompilationError
 from dbt.graph import ResourceTypeSelector
