+++ core/dbt/task/debug.py
@@ -21,7 +21,7 @@ import dbt.common.exceptions
 from dbt.adapters.factory import get_adapter, register_adapter
 from dbt.config import PartialProject, Project, Profile
 from dbt.config.renderer import DbtProjectYamlRenderer, ProfileRenderer
-from dbt.contracts.results import RunStatus
+from dbt.artifacts.results import RunStatus
 from dbt.clients.yaml_helper import load_yaml_text
 from dbt.links import ProfileConfigDocs
 from dbt.common.ui import green, red
