+++ core/dbt/task/docs/generate.py
@@ -15,8 +15,8 @@ from dbt.task.compile import CompileTask
 from dbt.adapters.factory import get_adapter
 from dbt.contracts.graph.nodes import ResultNode
 from dbt.contracts.graph.manifest import Manifest
-from dbt.contracts.results import (
-    NodeStatus,
+from dbt.artifacts.results import NodeStatus
+from dbt.artifacts.catalog import (
     TableMetadata,
     CatalogTable,
     CatalogResults,
