+++ core/dbt/task/run_operation.py
@@ -9,7 +9,8 @@ import dbt.common.exceptions
 from dbt.adapters.factory import get_adapter
 from dbt.contracts.files import FileHash
 from dbt.contracts.graph.nodes import HookNode
-from dbt.contracts.results import RunResultsArtifact, RunResult, RunStatus, TimingInfo
+from dbt.artifacts.results import RunStatus, TimingInfo
+from dbt.artifacts.run import RunResultsArtifact, RunResult
 from dbt.common.events.functions import fire_event
 from dbt.events.types import (
     LogDebugStackTrace,
