+++ core/dbt/task/runnable.py
@@ -15,14 +15,8 @@ from dbt.adapters.base import BaseRelation
 from dbt.adapters.factory import get_adapter
 from dbt.contracts.graph.manifest import WritableManifest
 from dbt.contracts.graph.nodes import ResultNode
-from dbt.contracts.results import (
-    NodeStatus,
-    RunExecutionResult,
-    RunningStatus,
-    RunResult,
-    RunStatus,
-    BaseResult,
-)
+from dbt.artifacts.results import NodeStatus, RunningStatus, RunStatus, BaseResult
+from dbt.artifacts.run import RunExecutionResult, RunResult
 from dbt.contracts.state import PreviousState
 from dbt.common.events.contextvars import log_contextvars, task_contextvars
 from dbt.common.events.functions import fire_event, warn_or_error
