+++ core/dbt/task/show.py
@@ -4,7 +4,7 @@ import time
 
 from dbt.context.providers import generate_runtime_model_context
 from dbt.contracts.graph.nodes import SeedNode
-from dbt.contracts.results import RunResult, RunStatus
+from dbt.artifacts.run import RunResult, RunStatus
 from dbt.common.events.base_types import EventLevel
 from dbt.common.events.functions import fire_event
 from dbt.common.events.types import Note
