+++ core/dbt/task/snapshot.py
@@ -6,7 +6,7 @@ from dbt.common.events.base_types import EventLevel
 from dbt.events.types import LogSnapshotResult
 from dbt.graph import ResourceTypeSelector
 from dbt.node_types import NodeType
-from dbt.contracts.results import NodeStatus
+from dbt.artifacts.results import NodeStatus
 from dbt.common.utils import cast_dict_to_dict_of_strings
